+++ CHANGELOG.md
@@ -4,7 +4,7 @@ UNRELEASED
  * see https://github.com/clux/kube-rs/compare/0.55.0...master
  * `kube`: added `Api::default_namespaced` - #209 via #534
  * `kube`: added `config` feature - #533 via #535
- * `kube`: BREAKING: moved `discovery` module to `kube::api` and reworked module #523
+ * `kube`: BREAKING: moved `client::discovery` module to `kube::discovery` and reworked module #523
  * `kube`: `discovery`: added helpers for recommended resources / kinds #523
 
 0.55.0 / 2021-05-21